requirements.txt
Django >= 1.4, < 1.5
django-annoying >= 0.7.6
django-tastypie==0.9.16
